Fix expr().
Basically it's all about expr() method which looses it's parent class and Query_MySQL->expr() for example never know that it should return Expression_MySQL not Expression class instance. In case we have Connection (not PDO) then we can get that info from Connection, because Connection class is/should be the one responsible for that. If we don't have Connection, but only PDO, then wehave to make "smart" quess (anyway that's not worse than it's now).